beginning of restructuring client logic for ticks

+++ src/network/server/server_config.h
@@ -9,8 +9,8 @@ typedef struct {
   uint8_t      max_clients;
   uint8_t      channels;
 
-  uint8_t      tick_rate;
-  uint8_t      snapshot_rate;
+  double       tick_rate;
+  double       snapshot_rate;
 } Server_Config;
 
 Server_Config* get_server_config(void);
